Do all copying in kernel mode for TCP on Linux
Uses `splice(2)` for the TCP proxy on Linux so that proxied data does not
need to pass through userspace at all.
Falls back to the "slow path" (user-mode copying) in case of any error
setting up the required b pieces for in-kernel copying.

Since `splice(2)` requires one side to be pipe, this requires setting up
a pipe where data is copied to first, but `SPLICE_F_MOVE` is used to
hopefully optimize any additional copying (pages are moved instead of
copied).

Uses epoll in edge-triggered mode to know when data is ready for
reads/writes.

Also uses a buffer pool for the "slow path" instead of allocating a new
buffer for each connection.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>